--- Sources/Alchemy/Rune/Model/ModelQuery.swift
@@ -1,64 +1,0 @@
-import Foundation
-import NIO
-
-public class ModelQuery<M: Model>: Query {
-    public typealias NestedEagerLoads<R: Model> = (ModelQuery<R>) -> ModelQuery<R>
-    
-    /// Right now these only run when the query is finished with `getAll` or `getFirst`. If the user finishes
-    /// with a `get()` we don't know if/when the decode will happen and how to handle it. Potential ways of
-    /// doing it (call eager load @ the `.decode` level of a `DatabaseRow`, but too complicated for now).
-    private var eagerLoads: [([M]) -> EventLoopFuture<[M]>] = []
-
-    public func getAll() -> EventLoopFuture<[M]> {
-        self.get(["\(M.tableName).*"])
-            .flatMapThrowing { try $0.map { try $0.decode(M.self) } }
-            .flatMap { self.evaluateEagerLoads(for: $0) }
-    }
-
-    public func getFirst() -> EventLoopFuture<M?> {
-        self.first(["\(M.tableName).*"])
-            .flatMapThrowing { try $0?.decode(M.self) }
-            .flatMap { result -> EventLoopFuture<M?> in
-                if let result = result {
-                    return self.evaluateEagerLoads(for: [result]).map { $0.first }
-                } else {
-                    return Loop.future(value: nil)
-                }
-            }
-    }
-    
-    public func unwrapFirst(or error: Error? = nil) -> EventLoopFuture<M> {
-        self.getFirst()
-            .flatMapThrowing { try $0.unwrap(or: error ?? RuneError(info: "Unable to find first element.")) }
-    }
-
-    public func with<R: Relationship>(
-        _ relationshipKeyPath: KeyPath<M, R>,
-        db: Database = DB.default,
-        nested: @escaping NestedEagerLoads<R.To.Value> = { $0 }
-    ) -> ModelQuery<M> where R.From == M {
-        self.eagerLoads.append { results in
-            // If there are no results, don't need to eager load.
-            guard let firstResult = results.first else {
-                return Loop.future(value: [])
-            }
-
-            return firstResult[keyPath: relationshipKeyPath]
-                .load(
-                    results,
-                    with: nested,
-                    from: relationshipKeyPath
-                )
-        }
-        
-        return self
-    }
-
-    /// Evaluate all of the eager loads, sequentially.
-    private func evaluateEagerLoads(for models: [M]) -> EventLoopFuture<[M]> {
-        self.eagerLoads
-            .reduce(Loop.future(value: models)) { future, eagerLoad in
-                future.flatMap { eagerLoad($0) }
-            }
-    }
-}
